parte de solicitud de viaje del lado del cliente en la base de datos de firebase

+++ app/src/main/java/com/equipo/ubertaxi/models/FCMResponse.java
@@ -4,7 +4,7 @@ package com.equipo.ubertaxi.models;
 import java.util.ArrayList;
 
 public class FCMResponse {
-    private int multicast_id;
+    private long multicast_id;
     private int success;
     private int failure;
     private int canonical_ids;
@@ -18,7 +18,7 @@ public class FCMResponse {
         this.results = results;
     }
 
-    public int getMulticast_id() {
+    public long getMulticast_id() {
         return multicast_id;
     }
